requirement.txt
cassandra-driver==3.25.0
click==8.0.1
data==0.4
decorator==5.0.9
Flask==2.0.1
funcsigs==1.0.2
geomet==0.2.1.post1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.1
pymongo==3.11.4
six==1.16.0
Werkzeug==2.0.1
